CXO-2951
Attempt to merge master into the UPL branch of 9.1 protocols. Everything seems to compile.
TODO: fix JUnit tests, make sure everything also works functionally.

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-us/src/main/java/com/elster/us/protocolimplv2/sel/profiles/LDPData.java
@@ -1,29 +1,33 @@
 package com.elster.us.protocolimplv2.sel.profiles;
 
-import java.util.List;
+import com.elster.us.protocolimplv2.sel.profiles.structure.LDPError;
+import com.elster.us.protocolimplv2.sel.profiles.structure.LPData;
+import com.elster.us.protocolimplv2.sel.profiles.structure.MeterConfiguration;
+import com.elster.us.protocolimplv2.sel.profiles.structure.MeterStatus;
+import com.elster.us.protocolimplv2.sel.profiles.structure.SERData;
 
-import com.elster.us.protocolimplv2.sel.profiles.structure.*;
+import java.util.List;
 
 
 public class LDPData {
-  
+
   private MeterConfiguration meterConfig;
-  private float[] presentValues;
+  private Long[] presentValues;
   private MeterStatus meterStatus;
   private List<LPData> lpData;
   private List<SERData> serData;
   private List<LDPError> ldpError;
-     
+
   public MeterConfiguration getMeterConfig() {
     return meterConfig;
   }
   public void setMeterConfig(MeterConfiguration meterConfig) {
     this.meterConfig = meterConfig;
   }
-  public float[] getPresentValues() {
+  public Long[] getPresentValues() {
     return presentValues;
   }
-  public void setPresentValues(float[] presentValues) {
+  public void setPresentValues(Long[] presentValues) {
     this.presentValues = presentValues;
   }
   public MeterStatus getMeterStatus() {
